refactor(!): deprecate top-level action setters
